refactor: move E2E tests from instancetype to cloudprovider
Move all tests that provision VMs (ExpectProvisionedAndWaitForPromises)
from pkg/providers/instancetype/suite_test.go to
pkg/cloudprovider/suite_instancetype_e2e_test.go.

instancetype/suite_test.go now only contains tests that call
InstanceTypesProvider.List() directly (LocalDNS filtering, SKU
filtering, GPU filtering, encryption at host, MaxPods) and pure
algorithm tests (FindMaxEphemeralSizeGBAndPlacement, KubeReservedResources).
